Add support for %VAR% syntax
To implement this, i added control flow simulation as variables are expanded instead of stored in the actual output executable. Currently, this control flow simulation only supports jumping to labels as other control flows are NOT implemented. I then added the actual expanding in the expand_string function of expand_vars.